requirements.txt
numpy
pandas
Pillow
scipy
pyimagej
scyjava
jpype1
